+++ data_structures/VertexEdge.h
@@ -67,19 +67,19 @@ public:
     Edge(Vertex *orig, Vertex *dest, double w, const string& service);
 
     Vertex * getDest() const;
-    double getWeight() const;
+    int getWeight() const;
     string getService() const;
     bool isSelected() const;
     Vertex * getOrig() const;
     Edge *getReverse() const;
-    double getFlow() const;
+    int getFlow() const;
 
     void setSelected(bool selected);
     void setReverse(Edge *reverse);
     void setFlow(double flow);
 protected:
     Vertex * dest; // destination vertex
-    double weight; // edge weight, can also be used for capacity
+    int weight; // edge weight, can also be used for capacity
     string service;
 
     // auxiliary fields
@@ -89,7 +89,7 @@ protected:
     Vertex *orig;
     Edge *reverse = nullptr;
 
-    double flow; // for flow-related problems
+    int flow; // for flow-related problems
 };
 
 #endif /* DA_TP_CLASSES_VERTEX_EDGE */
